+++ test/ReactViews/Workbench/Controls/IdealZoomSpec.tsx
@@ -9,6 +9,12 @@ import ViewingControls from "../../../../lib/ReactViews/Workbench/Controls/Viewi
 import ViewState from "../../../../lib/ReactViewModels/ViewState";
 
 import CesiumMath from "terriajs-cesium/Source/Core/Math";
+import {
+  IdealZoomTraits,
+  InitialCameraTraits,
+  LookAtTraits
+} from "../../../../lib/Traits/TraitsClasses/MappableTraits";
+import createStratumInstance from "../../../../lib/Models/Definition/createStratumInstance";
 
 describe("Ideal Zoom", function() {
   let terria: Terria;
@@ -59,8 +65,8 @@ describe("Ideal Zoom", function() {
     expect(theCameraView.rectangle?.west).toBeCloseTo(rectangle.west, 6);
   });
 
-  it("should customise camera view if the given parameters are valid.", async function() {
-    const idealZoom = {
+  it("should customise camera view if the given lookAt parameters are valid.", async function() {
+    const lookAtValues = {
       targetLongitude: 150.60832,
       targetLatitude: -34.19483,
       targetHeight: 200,
@@ -68,7 +74,11 @@ describe("Ideal Zoom", function() {
       pitch: 15,
       range: 200
     };
+    const lookAt = createStratumInstance(LookAtTraits, lookAtValues);
+    const idealZoom = createStratumInstance(IdealZoomTraits);
+    idealZoom.lookAt = lookAt;
     theItem.setTrait("definition", "idealZoom", idealZoom);
+
     await theItem.loadMapItems();
     act(() => {
       testRenderer = TestRenderer.create(
@@ -126,7 +136,7 @@ describe("Ideal Zoom", function() {
     theItem.setTrait("definition", "rectangle", wrappedRectangle);
 
     // But using home camera rule will be orverridden by idealZoom.
-    const idealZoom = {
+    const lookAtValues = {
       targetLongitude: 150.60832,
       targetLatitude: -34.19483,
       targetHeight: 200,
@@ -134,6 +144,9 @@ describe("Ideal Zoom", function() {
       pitch: 15,
       range: 200
     };
+    const lookAt = createStratumInstance(LookAtTraits, lookAtValues);
+    const idealZoom = createStratumInstance(IdealZoomTraits);
+    idealZoom.lookAt = lookAt;
     theItem.setTrait("definition", "idealZoom", idealZoom);
 
     await theItem.loadMapItems();
@@ -182,15 +195,19 @@ describe("Ideal Zoom", function() {
     expect(theCameraView.rectangle?.west).toBeCloseTo(rectangle.west, 6);
   });
 
-  it("should use default camera view if the given parameters are invalid.", async function() {
-    const idealZoom = {
+  it("should use default camera view if the given lookAt parameters are invalid.", async function() {
+    const lookAtValues = {
       targetLongitude: undefined,
-      targetLatitude: undefined,
+      targetLatitude: -34.19483,
       targetHeight: 200,
       heading: 180,
       pitch: 15,
       range: 200
     };
+    const lookAt = createStratumInstance(LookAtTraits, lookAtValues);
+    const idealZoom = createStratumInstance(IdealZoomTraits);
+    idealZoom.lookAt = lookAt;
+
     theItem.setTrait("definition", "idealZoom", idealZoom);
     await theItem.loadMapItems();
 
@@ -208,8 +225,8 @@ describe("Ideal Zoom", function() {
     expect(theCameraView.up).toBe(undefined);
   });
 
-  it("should use initialCamera if the given parameters are complete.", async function() {
-    const initialCamera = {
+  it("should custimise camera view if the given initialCamera parameters are complete.", async function() {
+    const initialCameraValues = {
       west: 143.85665964592238,
       south: -37.5588985189224,
       east: 143.85932639124115,
@@ -231,7 +248,14 @@ describe("Ideal Zoom", function() {
       }
     };
 
-    theItem.setTrait("definition", "initialCamera", initialCamera);
+    const initialCamera = createStratumInstance(
+      InitialCameraTraits,
+      initialCameraValues
+    );
+    const idealZoom = createStratumInstance(IdealZoomTraits);
+    idealZoom.initialCamera = initialCamera;
+
+    theItem.setTrait("definition", "idealZoom", idealZoom);
     await theItem.loadMapItems();
     act(() => {
       testRenderer = TestRenderer.create(
@@ -243,46 +267,55 @@ describe("Ideal Zoom", function() {
     const theCameraView = terria.currentViewer.getCurrentCameraView();
 
     expect(theCameraView.direction?.x).toBeCloseTo(
-      initialCamera.direction.x,
+      initialCameraValues.direction.x,
       6
     );
     expect(theCameraView.direction?.y).toBeCloseTo(
-      initialCamera.direction.y,
+      initialCameraValues.direction.y,
       6
     );
     expect(theCameraView.direction?.z).toBeCloseTo(
-      initialCamera.direction.z,
+      initialCameraValues.direction.z,
       6
     );
 
-    expect(theCameraView.position?.x).toBeCloseTo(initialCamera.position.x, 6);
-    expect(theCameraView.position?.y).toBeCloseTo(initialCamera.position.y, 6);
-    expect(theCameraView.position?.z).toBeCloseTo(initialCamera.position.z, 6);
+    expect(theCameraView.position?.x).toBeCloseTo(
+      initialCameraValues.position.x,
+      6
+    );
+    expect(theCameraView.position?.y).toBeCloseTo(
+      initialCameraValues.position.y,
+      6
+    );
+    expect(theCameraView.position?.z).toBeCloseTo(
+      initialCameraValues.position.z,
+      6
+    );
 
-    expect(theCameraView.up?.x).toBeCloseTo(initialCamera.up.x, 6);
-    expect(theCameraView.up?.y).toBeCloseTo(initialCamera.up.y, 6);
-    expect(theCameraView.up?.z).toBeCloseTo(initialCamera.up.z, 6);
+    expect(theCameraView.up?.x).toBeCloseTo(initialCameraValues.up.x, 6);
+    expect(theCameraView.up?.y).toBeCloseTo(initialCameraValues.up.y, 6);
+    expect(theCameraView.up?.z).toBeCloseTo(initialCameraValues.up.z, 6);
 
     expect(theCameraView.rectangle?.east).toBeCloseTo(
-      CesiumMath.toRadians(initialCamera.east),
+      CesiumMath.toRadians(initialCameraValues.east),
       6
     );
     expect(theCameraView.rectangle?.north).toBeCloseTo(
-      CesiumMath.toRadians(initialCamera.north),
+      CesiumMath.toRadians(initialCameraValues.north),
       6
     );
     expect(theCameraView.rectangle?.south).toBeCloseTo(
-      CesiumMath.toRadians(initialCamera.south),
+      CesiumMath.toRadians(initialCameraValues.south),
       6
     );
     expect(theCameraView.rectangle?.west).toBeCloseTo(
-      CesiumMath.toRadians(initialCamera.west),
+      CesiumMath.toRadians(initialCameraValues.west),
       6
     );
   });
 
   it("should use given rectangle if any other values are missing in initialCamera.", async function() {
-    const initialCamera = {
+    const initialCameraValues = {
       west: 143.85665964592238,
       south: -37.5588985189224,
       east: 143.85932639124115,
@@ -290,7 +323,7 @@ describe("Ideal Zoom", function() {
       position: {
         x: -4088564.111098504,
         y: 2985823.720726511,
-        z: undefined
+        z: -3867066.0705771768
       },
       direction: {
         x: 0.25654239033528903,
@@ -300,11 +333,18 @@ describe("Ideal Zoom", function() {
       up: {
         x: -0.6168995401836412,
         y: 0.5624968267277541,
-        z: -0.5504836757274634
+        z: undefined
       }
     };
 
-    theItem.setTrait("definition", "initialCamera", initialCamera);
+    const initialCamera = createStratumInstance(
+      InitialCameraTraits,
+      initialCameraValues
+    );
+    const idealZoom = createStratumInstance(IdealZoomTraits);
+    idealZoom.initialCamera = initialCamera;
+
+    theItem.setTrait("definition", "idealZoom", idealZoom);
     await theItem.loadMapItems();
     act(() => {
       testRenderer = TestRenderer.create(
@@ -320,25 +360,25 @@ describe("Ideal Zoom", function() {
     expect(theCameraView.up).toBe(undefined);
 
     expect(theCameraView.rectangle?.east).toBeCloseTo(
-      CesiumMath.toRadians(initialCamera.east),
+      CesiumMath.toRadians(initialCameraValues.east),
       6
     );
     expect(theCameraView.rectangle?.north).toBeCloseTo(
-      CesiumMath.toRadians(initialCamera.north),
+      CesiumMath.toRadians(initialCameraValues.north),
       6
     );
     expect(theCameraView.rectangle?.south).toBeCloseTo(
-      CesiumMath.toRadians(initialCamera.south),
+      CesiumMath.toRadians(initialCameraValues.south),
       6
     );
     expect(theCameraView.rectangle?.west).toBeCloseTo(
-      CesiumMath.toRadians(initialCamera.west),
+      CesiumMath.toRadians(initialCameraValues.west),
       6
     );
   });
 
   it("should use default camera view if missing any required rectagle parameters in initialCamera.", async function() {
-    const initialCamera = {
+    const initialCameraValues = {
       west: undefined,
       south: -37.5588985189224,
       east: 143.85932639124115,
@@ -360,7 +400,14 @@ describe("Ideal Zoom", function() {
       }
     };
 
-    theItem.setTrait("definition", "initialCamera", initialCamera);
+    const initialCamera = createStratumInstance(
+      InitialCameraTraits,
+      initialCameraValues
+    );
+    const idealZoom = createStratumInstance(IdealZoomTraits);
+    idealZoom.initialCamera = initialCamera;
+
+    theItem.setTrait("definition", "idealZoom", idealZoom);
     await theItem.loadMapItems();
     act(() => {
       testRenderer = TestRenderer.create(
@@ -381,6 +428,95 @@ describe("Ideal Zoom", function() {
       north: 1.5707963267948966,
       south: -1.5707963267948966,
       west: -3.141592653589793
+    };
+    expect(theCameraView.rectangle?.east).toBeCloseTo(rectangle.east, 6);
+    expect(theCameraView.rectangle?.north).toBeCloseTo(rectangle.north, 6);
+    expect(theCameraView.rectangle?.south).toBeCloseTo(rectangle.south, 6);
+    expect(theCameraView.rectangle?.west).toBeCloseTo(rectangle.west, 6);
+  });
+
+  it("should custimise camera view based on the given lookAt parameters when initialCamera parameters also exist.", async function() {
+    const lookAtValues = {
+      targetLongitude: 150.60832,
+      targetLatitude: -34.19483,
+      targetHeight: 200,
+      heading: 180,
+      pitch: 15,
+      range: 200
+    };
+    const lookAt = createStratumInstance(LookAtTraits, lookAtValues);
+
+    const idealZoom = createStratumInstance(IdealZoomTraits);
+    idealZoom.lookAt = lookAt;
+    const initialCameraValues = {
+      west: 143.85665964592238,
+      south: -37.5588985189224,
+      east: 143.85932639124115,
+      north: -37.55761610087383,
+      position: {
+        x: -4088564.111098504,
+        y: 2985823.720726511,
+        z: -3867066.0705771768
+      },
+      direction: {
+        x: 0.25654239033528903,
+        y: 0.8049473516085732,
+        z: 0.5350194044138963
+      },
+      up: {
+        x: -0.6168995401836412,
+        y: 0.5624968267277541,
+        z: -0.5504836757274634
+      }
+    };
+
+    const initialCamera = createStratumInstance(
+      InitialCameraTraits,
+      initialCameraValues
+    );
+    idealZoom.initialCamera = initialCamera;
+
+    theItem.setTrait("definition", "idealZoom", idealZoom);
+
+    await theItem.loadMapItems();
+    act(() => {
+      testRenderer = TestRenderer.create(
+        <ViewingControls item={theItem} viewState={viewState} />
+      );
+    });
+
+    testRenderer.root.findAllByType("button")[0].props.onClick();
+    const theCameraView = terria.currentViewer.getCurrentCameraView();
+
+    const directionX = 0.6595071845691584;
+    const directionY = -0.37148703631220265;
+    const directionZ = -0.6534888333809832;
+    expect(theCameraView.direction?.x).toBeCloseTo(directionX, 6);
+    expect(theCameraView.direction?.y).toBeCloseTo(directionY, 6);
+    expect(theCameraView.direction?.z).toBeCloseTo(directionZ, 6);
+
+    const positionX = -4601657.155837906;
+    const positionY = 2592020.25230978;
+    const positionZ = -3564324.3086873693;
+    expect(theCameraView.position?.x).toBeCloseTo(positionX, 6);
+    expect(theCameraView.position?.y).toBeCloseTo(positionY, 6);
+    expect(theCameraView.position?.z).toBeCloseTo(positionZ, 6);
+
+    const upX = -0.5693750748843667;
+    const upY = 0.3207174448857751;
+    const upZ = -0.7569361562551771;
+    expect(theCameraView.up?.x).toBeCloseTo(upX, 6);
+    expect(theCameraView.up?.y).toBeCloseTo(upY, 6);
+    expect(theCameraView.up?.z).toBeCloseTo(upZ, 6);
+
+    // The rectangle values are calculated from the given idealZoom parameters,
+    // i.e., the customised camera's direction, position and up parameters.
+    // A 2D viewer will zoom to this rectangle only.
+    const rectangle = {
+      east: 2.6286473947864932,
+      north: -0.5967760407704654,
+      south: -0.596848700549,
+      west: 2.628574735007959
     };
     expect(theCameraView.rectangle?.east).toBeCloseTo(rectangle.east, 6);
     expect(theCameraView.rectangle?.north).toBeCloseTo(rectangle.north, 6);